refactor(context_compressor): extract summary kernel mixin
